change annual_rate_of_return from float to decimal, refactored calculate_return/1 function to be more clear about what the equation is doing

+++ lib/next_steps/resources/return_input.ex
@@ -14,7 +14,7 @@ defmodule NextSteps.ReturnInput do
       allow_nil?(false)
     end
 
-    attribute :annual_rate_of_return, :float do
+    attribute :annual_rate_of_return, :decimal do
       allow_nil?(false)
     end
